Barriers to physical activity

Participants responses to the questionnaire assessing barriers to physical activity 12 months after the end of the cardiac rehabilitation program.

Usage

BARRIERS

Format

BARRIERS

A data frame with 77 rows and 12 columns:

patient

Participant's ID.

num_visit

1 = visit at the end of the cardiac rehabilitation program; 2 = visit at 6 months post program; 3 = visit at 12 months post program.

trop_vieux

"Too old"; 1 = "yes"; 0 = "no".

manque_interet

"Lack of interest"; 1 = "yes"; 0 = "no".

effort_import_fatig

"Heavy effort / too tired"; 1 = "yes"; 0 = "no".

manque_temps

"Lack of time"; 1 = "yes"; 0 = "no".

meteo_defavorable

"Unfavourable weather"; 1 = "yes"; 0 = "no".

deplacements_diff

"Difficulty to move"; 1 = "yes"; 0 = "no".

cout_trop_eleve

"Too costly"; 1 = "yes"; 0 = "no".

crainte_blessures_douleurs

"Fear of injury / pain"; 1 = "yes"; 0 = "no".

isolement_faible_RS

"Social isolation / weak social network"; 1 = "yes"; 0 = "no".

autres

"Other reasons".

autre_precision_obstacle

"Other details about barriers".

Source

APA&Co project