Physical, clinical, and functional characteristics at the inclusion visit

Participants data relating to the physical, clinical, and functional profiles at the end of the cardiac rehabilitation program.

Usage

INCLUSION

Format

INCLUSION

A data frame with 83 rows and 8 columns:

patient

Participant's ID.

age

Age in years.

sex

1 = male; 2 = female.

height

Height in cm.

weight

Weight in kg.

angioplasty

1 = Has had an angioplasty; 0 = Has not had an angioplasty.

bypas

1 = Has had a bypass surgery; 0 = Has not had a bypass surgery.

DIST_6WT_M0

6-min walking test distance in m.

Source

APA&Co project