IPAQ-SF

Participants responses to the IPAQ-SF at 0, 6 and 12 months after the end of the cardiac rehabilitation program.

Usage

IPAQ

Format

IPAQ

A data frame with 243 rows and 13 columns:

patient

Participant's ID.

num_visit

1 = visit at the end of the cardiac rehabilitation program; 2 = visit at 6 months post program; 3 = visit at 12 months post program.

bloc1_q1

1 = The participant engaged in vigorous physical activity (>= 10 min bouts) during the last 7 days; 0 = The participant did not engage in vigorous physical activity (>= 10 min bouts) during the last 7 days.

bloc2_q2

Number of days with vigorous physical activity (>= 10 min bouts) during the last 7 days.

total_hours_heavy

Total time of vigorous physical activity (>= 10 min bouts) during the last 7 days (hours part); 999 = Did not answer.

total_minutes_heavy

Total time of vigorous physical activity (>= 10 min bouts) during the last 7 days (minutes part); 999 = Did not answer.

bloc2_q1

1 = The participant engaged in moderate physical activity (>= 10 min bouts) during the last 7 days; 0 = The participant did not engage in moderate physical activity (>= 10 min bouts) during the last 7 days.

bloc2_q2

Number of days with moderate physical activity (>= 10 min bouts) during the last 7 days.

total_hours_moderate

Total time of moderate physical activity (>= 10 min bouts) during the last 7 days (hours part); 999 = Did not answer.

total_minutes_moderate

Total time of moderate physical activity (>= 10 min bouts) during the last 7 days (minutes part); 999 = Did not answer.

bloc3_q1

1 = The participant engaged in walking activity (>= 10 min bouts) during the last 7 days; 0 = The participant did not engage in walking activity (>= 10 min bouts) during the last 7 days.

bloc3_q2

Number of days with walking activity (>= 10 min bouts) during the last 7 days.

bouts_walk_7days

Number of (consecutive or not) walking periods of 10 min during the last 7 days; 9999 = Did not answer.

Source

APA&Co project