Get a set of figures and tables for analysing the change in the variable of interest

Get a set of figures and tables for analysing the change in the variable of interest

Usage

analyse_change(
  data,
  id,
  x,
  y,
  rain_side = NULL,
  nudge_y = NULL,
  color_fill = NULL,
  color_stat = NULL,
  labs_1x = NULL,
  labs_1y = NULL,
  labs_2x = NULL,
  labs_2y = NULL,
  labs_3x = NULL,
  labs_3y = NULL,
  labs_4x = NULL,
  labs_4y = NULL,
  labs_5x = NULL,
  labs_5y = NULL,
  labs_6x = NULL,
  labs_6y = NULL
)

Arguments

data

A data frame.

id

A character value to indicate the name of the variable designating the identities of the observations.

x

A character value to indicate the name of a factor variable (2 levels) to be plotted on the X axis.

y

A character value to indicate the name of a numeric variable to be plotted on the Y axis.

rain_side

A character value to indicate how you want the rainclouds displayed, right ("r"), left ("l") or flanking ("f"), for a 1-by-1 flanking raincloud use ("f1x1") and for a 2-by-2 use ("f2x2").

nudge_y

A numeric value to vertically move the labels of the graphic showing the shift function.

color_fill

A character value to set the filling color.

color_stat

A character value to set the color of the statistical summaries.

labs_1x

A character string to name the X axis of the 1rst graphic (marginal distributions).

labs_1y

A character string to name the Y axis of the 1rst graphic (marginal distributions).

labs_2x

A character string to name the X axis of the 2nd graphic (pairwise differences).

labs_2y

A character string to name the Y axis of the 2nd graphic (pairwise differences).

labs_3x

A character string to name the X axis of the 3rd graphic (bivariate relationship).

labs_3y

A character string to name the Y axis of the 3rd graphic (bivariate relationship).

labs_4x

A character string to name the X axis of the 4th graphic (quantile shifts).

labs_4y

A character string to name the Y axis of the 4th graphic (quantile shifts).

labs_5x

A character string to name the X axis of the 5th graphic (shift function).

labs_5y

A character string to name the Y axis of the 5th graphic (shift function).

labs_6x

A character string to name the X axis of the 6th graphic (difference asymmetry function).

labs_6y

A character string to name the Y axis of the 6th graphic (difference asymmetry function).

Value

A list of objects (one figure, and three tables: the estimates of the quantiles of the individual changes, the shift function, and the difference asymmetry function).