Make a polished raincloud plot for visualising the change

Make a polished raincloud plot for visualising the change

Usage

view_rainclouds(
  data,
  id,
  x,
  y,
  rain_side = NULL,
  color_fill = NULL,
  color_stat = NULL,
  labs_x = NULL,
  labs_y = NULL
)

Arguments

data

A data frame.

id

A character value to indicate the name of the variable designating the identities of the observations.

x

A character value to indicate the name of a factor variable (2 levels) to be plotted on the X axis.

y

A character value to indicate the name of a numeric variable to be plotted on the Y axis.

rain_side

A character value to indicate how you want the rainclouds displayed, right ("r"), left ("l") or flanking ("f"), for a 1-by-1 flanking raincloud use ("f1x1") and for a 2-by-2 use ("f2x2").

color_fill

A character value to set the filling color.

color_stat

A character value to set the color of the statistical summaries.

labs_x

A character string to name the X axis.

labs_y

A character string to name the Y axis.

Value

A ggplot object.